Refactored home page
